implemented songs download, appbar functions refactoring, need to fix bug on load of downloaded songs
